+++ src/features/editor/lib/encodeBase64.ts
