+++ src/features/editor/ui/MethodEditor.tsx
@@ -5,10 +5,10 @@ import {
   Select,
   SelectChangeEvent,
 } from '@mui/material';
-import { ICRUD, IMethodProps } from '../model/methodManagerModel';
+import { ICRUD, IMethodProps } from '../model/methodEditorModel';
 import { FC } from 'react';
 
-const MethodManager: FC<IMethodProps> = ({ method, setMethod }) => {
+const MethodEditor: FC<IMethodProps> = ({ method, setMethod }) => {
   const handleChange = (e: SelectChangeEvent<string>): void => {
     const { value } = e?.target || {};
 
@@ -39,4 +39,4 @@ const MethodManager: FC<IMethodProps> = ({ method, setMethod }) => {
   );
 };
 
-export default MethodManager;
+export default MethodEditor;
