+++ src/features/editor/ui/UrlEditor.tsx
@@ -1,12 +1,12 @@
 import { Box, TextField } from '@mui/material';
-import { IUrlProps } from '../model/urlManagerModel';
-import { ICRUD } from '../model/methodManagerModel';
-import { IHeadersData } from '../model/headerManagerModel';
-import { IBodyData } from '../model/bodyManagerModel';
+import { IUrlProps } from '../model/urlEditorModel';
+import { ICRUD } from '../model/methodEditorModel';
+import { IHeadersData } from '../model/headersEditorModel';
 import { FC } from 'react';
-import { encodeBase64 } from '../lib/encodeBase64';
+import { IBodyData } from '@widgets/restFullClient/model/bodyManagerModel';
+import { encodeBase64 } from '@features/editor/lib/encodeBase64';
 
-const UrlManager: FC<IUrlProps> = ({ url, setUrl, method, headers, body }) => {
+const UrlEditor: FC<IUrlProps> = ({ url, setUrl, method, headers, body }) => {
   const urlFormation = (
     method: ICRUD,
     headers: IHeadersData,
@@ -51,4 +51,4 @@ const UrlManager: FC<IUrlProps> = ({ url, setUrl, method, headers, body }) => {
   );
 };
 
-export default UrlManager;
+export default UrlEditor;
