+++ src/features/editor/ui/index.ts
@@ -1,0 +1,5 @@
+import HeaderEditor from './HeaderEditor';
+import MethodEditor from './MethodEditor';
+import UrlEditor from './UrlEditor';
+
+export { HeaderEditor, MethodEditor, UrlEditor };
