+++ src/shared/tests/__tests__/UrlEditor.test.tsx
@@ -2,12 +2,12 @@ import '@testing-library/jest-dom';
 import { act, fireEvent, screen } from '@testing-library/react';
 import { renderWithProviders } from '../lib';
 import { IBodyData } from '@widgets/restFullClient/model/bodyManagerModel';
-import { ICRUD } from '@widgets/restFullClient/model/methodManagerModel';
-import { IUrlProps } from '@widgets/restFullClient/model/urlManagerModel';
-import UrlManager from '@widgets/restFullClient/ui/UrlManager';
-import { IHeadersData } from '@widgets/restFullClient/model/headerManagerModel';
+import { ICRUD } from '@features/editor/model/methodEditorModel';
+import { IUrlProps } from '@features/editor/model/urlEditorModel';
+import { IHeadersData } from '@features/editor/model/headersEditorModel';
+import { UrlEditor } from '@features/editor';
 
-describe('UrlManager component', () => {
+describe('UrlEditor component', () => {
   const defaultProps: IUrlProps = {
     url: '',
     setUrl: jest.fn(),
@@ -17,22 +17,22 @@ describe('UrlManager component', () => {
   };
 
   const renderComponent = (props: Partial<IUrlProps> = {}) => {
-    return renderWithProviders(<UrlManager {...defaultProps} {...props} />);
+    return renderWithProviders(<UrlEditor {...defaultProps} {...props} />);
   };
 
   /**
-   * Тест для компонента UrlManager.
-   * Проверяет, что компонент UrlManager рендерится.
+   * Тест для компонента UrlEditor.
+   * Проверяет, что компонент UrlEditor рендерится.
    *
    * @returns {void}
    */
-  it('renders UrlManager component', () => {
+  it('renders UrlEditor component', () => {
     renderComponent();
     expect(screen.getByLabelText('Endpoint URL')).toBeInTheDocument();
   });
 
   /**
-   * Тест для компонента UrlManager.
+   * Тест для компонента UrlEditor.
    * Проверяет, что функция setUrl вызывается при изменении ввода.
    *
    * @returns {void}
