+++ src/shared/tests/__tests__/requestHandler.test.ts
@@ -1,5 +1,5 @@
 import { sendRequest } from '@widgets/restFullClient/lib/requestHandler';
-import { ICRUD } from '@widgets/restFullClient/model/methodManagerModel';
+import { ICRUD } from '@features/editor/model/methodEditorModel';
 import { IRequestData } from '@widgets/restFullClient/model/requestHandlerModel';
 import axios from 'axios';
 import MockAdapter from 'axios-mock-adapter';
