+++ src/widgets/restFullClient/model/requestHandlerModel.ts
@@ -1,6 +1,6 @@
 import { IBodyData } from './bodyManagerModel';
-import { IHeadersData } from './headerManagerModel';
-import { ICRUD } from './methodManagerModel';
+import { IHeadersData } from '../../../features/editor/model/headersEditorModel';
+import { ICRUD } from '../../../features/editor/model/methodEditorModel';
 
 export interface IRequestData {
   method: ICRUD;
